Fix build with later WDKs
- Adds alias for GetProjectInfoForReference target to version.vcxproj
    Later kits seemed to have renamed the build target, and will fail without
    this alias target.
- Adds "/fd sha256" to signtool command line
    WDK 20344 and later require binaries signed with a SHA256 file digest, or
    the build outputs are deleted
- Disables warning 4061 - switch statement on enum types need to have a case for
    all values of the enumeration


- Cast enum types used as array indices to avoid bounds check complaint

+++ src/xenvbd/driver.c
@@ -301,7 +301,7 @@ DriverGetFeatureOverride(
 {
     BOOLEAN              Present = FALSE;
 
-    if (Feature < ARRAYSIZE(Driver.FeatureOverride)) {
+    if ((ULONG)Feature < ARRAYSIZE(Driver.FeatureOverride)) {
         Present = Driver.FeatureOverride[Feature].Present;
         *Value = Driver.FeatureOverride[Feature].Value;
     }
@@ -315,7 +315,7 @@ DriverGetFeatureName(
     IN  XENVBD_FEATURE  Feature
     )
 {
-    return (Feature < ARRAYSIZE(Driver.FeatureOverride)) ?
+    return ((ULONG)Feature < ARRAYSIZE(Driver.FeatureOverride)) ?
            Driver.FeatureOverride[Feature].Name :
            NULL;
 }
